Add files via upload
A tic_tac_toe player that is going to beat you 90 percent of the time, without having to worry about game theory. Tic-Tac-Toe is a solved game, with deterministic strategies established for each game state within the framework of game theory, However, a neural network can learn to beat you instantly without having to worry about the time-consuming process of analysing the opponent's optimal moves . The repository consists of several Python files: one writing the script for the game using the famous pygame module, the other simulating Tic-Tac-Toe games using a deterministic strategy to determine the best moves for each possible state. Additionally, there is a CSV file containing the game's states along with the optimal moves for each game state, and a Jupyter notebook that creates the neural network.